Reorganize interfaces to better distinct native api from skype-http api.
