+++ package.json
@@ -42,6 +42,7 @@
     "chai": "^3.5.0",
     "del": "^2.2.0",
     "fs-extra": "^0.30.0",
+    "glob": "^7.0.3",
     "mocha": "^2.4.5",
     "tslint": "^3.9.0",
     "typescript": "^1.8.10"
