+++ src/demo/tsconfig.json
@@ -1,19 +1,61 @@
 {
   "compilerOptions": {
     "declaration": true,
-    "target": "ES5",
     "module": "commonjs",
+    "noResolve": true,
     "outDir": "../../build/demo",
-    "sourceMap": true,
     "removeComments": true,
-    "noImplicitAny": true
+    "sourceMap": true,
+    "target": "ES5"
   },
   "files": [
+    "../../typings/globals/mocha/index.d.ts",
+    "../../typings/globals/node/index.d.ts",
     "../../typings/index.d.ts",
+    "../../typings/modules/big-integer/index.d.ts",
+    "../../typings/modules/bluebird/index.d.ts",
+    "../../typings/modules/chai/index.d.ts",
+    "../../typings/modules/cheerio/index.d.ts",
+    "../../typings/modules/es6-promise/index.d.ts",
+    "../../typings/modules/form-data/index.d.ts",
+    "../../typings/modules/incident/index.d.ts",
+    "../../typings/modules/js-sha256/index.d.ts",
+    "../../typings/modules/lodash/index.d.ts",
+    "../../typings/modules/request/index.d.ts",
+    "../lib/api-uri.ts",
+    "../lib/api.ts",
+    "../lib/api/accept-contact-request.ts",
+    "../lib/api/decline-contact-request.ts",
+    "../lib/api/get-contact.ts",
+    "../lib/api/get-contacts.ts",
+    "../lib/api/get-conversation.ts",
+    "../lib/api/get-conversations.ts",
+    "../lib/api/get-messages.ts",
+    "../lib/api/send-message.ts",
+    "../lib/api/set-status.ts",
+    "../lib/connect.ts",
     "../lib/consts.ts",
+    "../lib/contacts-uri.ts",
+    "../lib/interfaces/api/api.ts",
+    "../lib/interfaces/api/contact.ts",
+    "../lib/interfaces/api/context.ts",
+    "../lib/interfaces/api/conversation.ts",
+    "../lib/interfaces/api/events.ts",
+    "../lib/interfaces/api/resources.ts",
+    "../lib/interfaces/io.ts",
+    "../lib/interfaces/native-api/contact.ts",
+    "../lib/interfaces/native-api/conversation.ts",
+    "../lib/interfaces/native-api/events.ts",
+    "../lib/interfaces/native-api/message-resources.ts",
+    "../lib/interfaces/native-api/resources.ts",
+    "../lib/interfaces/utils.ts",
     "../lib/login.ts",
+    "../lib/messages-uri.ts",
     "../lib/polling/messages-poller.ts",
+    "../lib/request-io.ts",
+    "../lib/skype-http.ts",
     "../lib/utils.ts",
+    "../lib/utils/formatters.ts",
     "../lib/utils/hmac-sha256.ts",
     "main.ts"
   ]
