+++ src/lib/api/decline-contact-request.ts
@@ -1,11 +1,11 @@
 import * as Bluebird from "bluebird";
 import {Incident} from "incident";
 
-import {ApiContext} from "../interfaces/api-context";
+import {Context} from "../interfaces/api/context";
 import * as io from "../interfaces/io";
 import * as apiUri from "../api-uri";
 
-export function declineContactRequest (io: io.IO, apiContext: ApiContext, contactUsername: string): Bluebird<any> {
+export function declineContactRequest (io: io.IO, apiContext: Context, contactUsername: string): Bluebird<any> {
   return Bluebird
     .try(() => {
       const requestOptions: io.GetOptions = {
