+++ src/lib/interfaces/api/events.ts
@@ -1,0 +1,13 @@
+import {ParsedConversationId} from "./api";
+import {Resource} from "./resources";
+
+export interface EventMessage {
+  id: number;
+  type: string;
+  resourceType: string;
+  time: Date;
+  resourceLink: string;
+  resource: Resource;
+}
+
+
