+++ src/lib/interfaces/utils.ts
@@ -1,3 +1,7 @@
 export interface Dictionary<T> {
   [key: string]: T;
 }
+
+export type Nullable <T> = T;
+
+export interface EmptyObject {}
