+++ src/lib/tsconfig.json
@@ -1,36 +1,61 @@
 {
   "compilerOptions": {
     "declaration": true,
-    "target": "ES5",
     "module": "commonjs",
+    "noResolve": true,
     "outDir": "../../build/local",
-    "sourceMap": true,
     "removeComments": true,
-    "noImplicitAny": true
+    "sourceMap": true,
+    "target": "ES5"
   },
   "files": [
+    "../../typings/globals/mocha/index.d.ts",
+    "../../typings/globals/node/index.d.ts",
     "../../typings/index.d.ts",
+    "../../typings/modules/big-integer/index.d.ts",
+    "../../typings/modules/bluebird/index.d.ts",
+    "../../typings/modules/chai/index.d.ts",
+    "../../typings/modules/cheerio/index.d.ts",
+    "../../typings/modules/es6-promise/index.d.ts",
+    "../../typings/modules/form-data/index.d.ts",
+    "../../typings/modules/incident/index.d.ts",
+    "../../typings/modules/js-sha256/index.d.ts",
+    "../../typings/modules/lodash/index.d.ts",
+    "../../typings/modules/request/index.d.ts",
+    "api-uri.ts",
+    "api.ts",
     "api/accept-contact-request.ts",
+    "api/decline-contact-request.ts",
+    "api/get-contact.ts",
     "api/get-contacts.ts",
+    "api/get-conversation.ts",
+    "api/get-conversations.ts",
+    "api/get-messages.ts",
     "api/send-message.ts",
     "api/set-status.ts",
-    "interfaces/api.ts",
-    "interfaces/api-context.ts",
-    "interfaces/index.ts",
-    "interfaces/io.ts",
-    "interfaces/native-api.ts",
-    "interfaces/utils.ts",
-    "polling/messages-poller.ts",
-    "utils/formatters.ts",
-    "utils/hmac-sha256.ts",
-    "api.ts",
     "connect.ts",
-    "messages-uri.ts",
     "consts.ts",
+    "contacts-uri.ts",
+    "interfaces/api/api.ts",
+    "interfaces/api/contact.ts",
+    "interfaces/api/context.ts",
+    "interfaces/api/conversation.ts",
+    "interfaces/api/events.ts",
+    "interfaces/api/resources.ts",
     "interfaces/io.ts",
+    "interfaces/native-api/contact.ts",
+    "interfaces/native-api/conversation.ts",
+    "interfaces/native-api/events.ts",
+    "interfaces/native-api/message-resources.ts",
+    "interfaces/native-api/resources.ts",
+    "interfaces/utils.ts",
     "login.ts",
+    "messages-uri.ts",
+    "polling/messages-poller.ts",
     "request-io.ts",
     "skype-http.ts",
-    "utils.ts"
+    "utils.ts",
+    "utils/formatters.ts",
+    "utils/hmac-sha256.ts"
   ]
 }
