+++ src/test/tsconfig.json
@@ -1,32 +1,65 @@
 {
   "compilerOptions": {
     "declaration": true,
-    "target": "ES5",
     "module": "commonjs",
-    "outDir": "../../build/test",
-    "sourceMap": true,
+    "noResolve": true,
+    "outDir": "../../build/demo",
     "removeComments": true,
-    "noImplicitAny": true
+    "sourceMap": true,
+    "target": "ES5"
   },
   "files": [
+    "../../typings/globals/mocha/index.d.ts",
+    "../../typings/globals/node/index.d.ts",
     "../../typings/index.d.ts",
-    "../lib/interfaces/api-context.ts",
-    "../lib/interfaces/index.ts",
-    "../lib/interfaces/io.ts",
-    "../lib/interfaces/utils.ts",
-    "../lib/polling/messages-poller.ts",
-    "../lib/utils/hmac-sha256.ts",
+    "../../typings/modules/big-integer/index.d.ts",
+    "../../typings/modules/bluebird/index.d.ts",
+    "../../typings/modules/chai/index.d.ts",
+    "../../typings/modules/cheerio/index.d.ts",
+    "../../typings/modules/es6-promise/index.d.ts",
+    "../../typings/modules/form-data/index.d.ts",
+    "../../typings/modules/incident/index.d.ts",
+    "../../typings/modules/js-sha256/index.d.ts",
+    "../../typings/modules/lodash/index.d.ts",
+    "../../typings/modules/request/index.d.ts",
+    "../lib/api-uri.ts",
     "../lib/api.ts",
-    "../lib/messages-uri.ts",
-    "../lib/consts.ts",
-    "../lib/interfaces/io.ts",
-    "../lib/login.ts",
-    "../lib/request-io.ts",
+    "../lib/api/accept-contact-request.ts",
+    "../lib/api/decline-contact-request.ts",
+    "../lib/api/get-contact.ts",
+    "../lib/api/get-contacts.ts",
+    "../lib/api/get-conversation.ts",
+    "../lib/api/get-conversations.ts",
+    "../lib/api/get-messages.ts",
+    "../lib/api/send-message.ts",
+    "../lib/api/set-status.ts",
     "../lib/connect.ts",
+    "../lib/consts.ts",
+    "../lib/contacts-uri.ts",
+    "../lib/interfaces/api/api.ts",
+    "../lib/interfaces/api/contact.ts",
+    "../lib/interfaces/api/context.ts",
+    "../lib/interfaces/api/conversation.ts",
+    "../lib/interfaces/api/events.ts",
+    "../lib/interfaces/api/resources.ts",
+    "../lib/interfaces/io.ts",
+    "../lib/interfaces/native-api/contact.ts",
+    "../lib/interfaces/native-api/conversation.ts",
+    "../lib/interfaces/native-api/events.ts",
+    "../lib/interfaces/native-api/message-resources.ts",
+    "../lib/interfaces/native-api/resources.ts",
+    "../lib/interfaces/utils.ts",
+    "../lib/login.ts",
+    "../lib/messages-uri.ts",
+    "../lib/polling/messages-poller.ts",
+    "../lib/request-io.ts",
+    "../lib/skype-http.ts",
     "../lib/utils.ts",
+    "../lib/utils/formatters.ts",
+    "../lib/utils/hmac-sha256.ts",
     "end-to-end/config.ts",
     "end-to-end/skype-http.spec.ts",
-    "utils/hmac-sha256.spec.ts",
-    "utils.spec.ts"
+    "utils.spec.ts",
+    "utils/hmac-sha256.spec.ts"
   ]
 }
